+++ src/_view/app.c
@@ -2,6 +2,7 @@
 #include <view/app.h>
 #include <view/awaitable.h>
 #include <view/map.h>
+#include <view/view.h>
 #include <stdbool.h>
 #include <stdint.h>
 #define ER(code, str) case code: return str
@@ -64,6 +65,7 @@
 #define TYPECODE_NONE 6
 #define TYPECODE_CLASS 7
 #define TYPECODE_CLASSTYPES 8
+#define TYPECODE_LIST 9
 
 typedef struct _route_input route_input;
 typedef struct _app_parsers app_parsers;
@@ -383,10 +385,18 @@ static PyObject* query_parser(
                     verified = true; \
                 } break;
 
+static PyObject* cast_from_typecodes(
+    type_info** codes,
+    Py_ssize_t len,
+    PyObject* item,
+    PyObject* json_parser
+);
+
 static int verify_dict_typecodes(
     type_info** codes,
     Py_ssize_t len,
-    PyObject* dict
+    PyObject* dict,
+    PyObject* json_parser
 ) {
     if (!PyDict_Size(dict)) return 0;
     PyObject* iter = PyObject_GetIter(dict);
@@ -401,51 +411,61 @@ static int verify_dict_typecodes(
             return -1;
         }
 
-        bool verified = false;
-        for (Py_ssize_t i = 0; i < len; i++) {
-            if (verified) break;
-            switch (codes[i]->typecode) {
-            case TYPECODE_ANY: return 0;
-            case TYPECODE_NONE: {
-                if (value == Py_None) verified = true;
-                break;
-            }
-            case TYPECODE_STR: TC_VERIFY(PyUnicode_CheckExact);
-            case TYPECODE_INT: {
-                if (PyLong_CheckExact(value)) {
-                    verified = true;
-                }
-                break;
-            };
-            case TYPECODE_BOOL: TC_VERIFY(PyBool_Check);
-            case TYPECODE_FLOAT: TC_VERIFY(PyFloat_CheckExact);
-            case TYPECODE_DICT: {
-                if (PyObject_IsInstance(
-                    value,
-                    (PyObject*) &PyDict_Type
-                    )) {
-                    int res = verify_dict_typecodes(
-                        codes[i]->children,
-                        codes[i]->children_size,
-                        value
-                    );
-                    if (res == 0) verified = true;
-                    else if (res == -1) return -1;
-                    else return 1;
-                };
-                break;
-            };
-            case TYPECODE_CLASSTYPES:
-            default: Py_FatalError("invalid dict typecode");
-            };
+        value = cast_from_typecodes(
+            codes,
+            len,
+            value,
+            json_parser
+        );
+        if (!value) return -1;
+        if (PyDict_SetItem(
+            dict,
+            key,
+            value
+            ) < 0) {
+            Py_DECREF(iter);
+            return -1;
         }
-        if (!verified) return 1;
     }
 
     Py_DECREF(iter);
     if (PyErr_Occurred()) {
         PyErr_Print();
         return -1;
+    }
+
+    return 0;
+}
+
+static int verify_list_typecodes(
+    type_info** codes,
+    Py_ssize_t len,
+    PyObject* list,
+    PyObject* json_parser
+) {
+    Py_ssize_t list_len = PySequence_Size(list);
+    if (list_len == -1) return -1;
+    if (list_len == 0) return 0;
+
+    for (int i = 0; i < list_len; i++) {
+        PyObject* item = PyList_GET_ITEM(
+            list,
+            i
+        );
+
+        item = cast_from_typecodes(
+            codes,
+            len,
+            item,
+            json_parser
+        );
+
+        if (!item) return 1;
+        PyList_SET_ITEM(
+            list,
+            i,
+            item
+        );
     }
 
     return 0;
@@ -551,16 +571,19 @@ static PyObject* cast_from_typecodes(
             int res = verify_dict_typecodes(
                 ti->children,
                 ti->children_size,
-                obj
+                obj,
+                json_parser
             );
             if (res == -1) {
                 Py_DECREF(obj);
                 return NULL;
             }
+
             if (res == 1) {
                 Py_DECREF(obj);
-                return NULL;
+                break;
             }
+
             return obj;
         }
         case TYPECODE_CLASS: {
@@ -589,7 +612,6 @@ static PyObject* cast_from_typecodes(
 
             bool ok = true;
             for (Py_ssize_t i = 0; i < ti->children_size; i++) {
-
                 type_info* info = ti->children[i];
                 PyObject* got_item = PyDict_GetItem(
                     obj,
@@ -675,6 +697,54 @@ static PyObject* cast_from_typecodes(
 
             return built;
         }
+        case TYPECODE_LIST: {
+            PyObject* list;
+            if (Py_IS_TYPE(
+                item,
+                &PyList_Type
+                )) {
+                Py_INCREF(item);
+                list = item;
+            }
+            else {
+                list = PyObject_Vectorcall(
+                    json_parser,
+                    (PyObject*[]) { item },
+                    1,
+                    NULL
+                );
+
+                if (!list) {
+                    PyErr_Clear();
+                    break;
+                }
+
+                if (!Py_IS_TYPE(
+                    list,
+                    &PyList_Type
+                    )) {
+                    break;
+                }
+            }
+
+            int res = verify_list_typecodes(
+                ti->children,
+                ti->children_size,
+                list,
+                json_parser
+            );
+            if (res == -1) {
+                Py_DECREF(list);
+                return NULL;
+            }
+
+            if (res == 1) {
+                Py_DECREF(list);
+                break;
+            }
+
+            return list;
+        }
         case TYPECODE_CLASSTYPES:
         default: {
             fprintf(
@@ -682,7 +752,7 @@ static PyObject* cast_from_typecodes(
                 "got bad typecode in cast_from_typecodes: %d\n",
                 ti->typecode
             );
-            Py_FatalError("invalid typecode");
+            VIEW_FATAL("invalid typecode");
         }
         }
     }
@@ -737,7 +807,6 @@ static PyObject** json_parser(
             inp->is_body ? obj : query,
             inp->name
         );
-
         PyObject* item = cast_from_typecodes(
             inp->types,
             inp->types_size,
@@ -835,7 +904,7 @@ static int send_raw_text(
     PyObject* send,
     int status,
     const char* res_str,
-    PyObject* headers /* may be NULL */
+    PyObject* headers     /* may be NULL */
 ) {
     PyObject* coro;
     PyObject* send_dict;
@@ -1152,7 +1221,8 @@ static const char* get_err_str(int status) {
     );
     }
 
-    Py_FatalError("got bad status code");
+    VIEW_FATAL("got bad status code");
+    return NULL;
 }
 
 static int find_result_for(
@@ -3356,7 +3426,7 @@ int load_parts(ViewApp* app, map* routes, PyObject* parts, route* r) {
             rt = transport;
         } else {
             app->has_path_params = true;
-            if (!rt) Py_FatalError("first path param was part");
+            if (!rt) VIEW_FATAL("first path param was part");
             if (index == size) {
                 rt->r = r;
                 set_r = true;