+++ src/view/__init__.py
@@ -8,13 +8,16 @@ GitHub: https://github.com/zerointensity/view.py
 try:
     import _view
 except ImportError as e:
-    raise ImportError("_view has not been built, did you forget to compile it?") from e
+    raise ImportError(
+        "_view has not been built, did you forget to compile it?"
+    ) from e
 
 from . import _codec
 from .__about__ import __license__, __version__
 from .app import *
 from .components import *
 from .exceptions import *
+from .logging import *
 from .response import *
 from .routing import *
 from .util import *